Initial Commit of WRI-tidy notebook
The long-term goal of this notebook is to demonstrate the proper onboarding of WRI Power Plant data into the data pipeline architecture.  In this initial implementation we read the raw data from an S3 bucket and perform a simple Pandas melt operation to tidy the data into two tables: wri_plants and wri_annual_gwh.