Update pudl_ingestion_sample.ipynb outputs
Run notebook fully so that outputs reflect code.  Note that vcaldeira is the owner of Y95_AL and MichaelTiemannOSC is the owner of Y96_AL (mistakenly placed in the epacems_y95_al schema).  This notebook should really initialize the schema and table spaces for the user running them.  Or should create per-user schema spaces for demo purposes.