Aqueduct.ipynb initial commit
This initial commit is very incomplete.  We read a geotif file from a remote WRI S3 bucket (so far so good) and transform it to a resampled file we store in notebook-local storage (not good).  We really need to figure out what is the data that needs to be stored in what way into trino.